+++ SourceryRuntime/Sources/Method.swift
@@ -205,6 +205,12 @@ public typealias SourceryMethod = Method
         return shortName.hasSuffix(">")
     }
 
+    // sourcery: skipEquality, skipDescription
+    /// Whether method is optional (in an Objective-C protocol)
+    public var isOptional: Bool {
+        return attributes[Attribute.Identifier.optional.name] != nil
+    }
+
     /// Annotations, that were created with // sourcery: annotation1, other = "annotation value", alterantive = 2
     public let annotations: [String: NSObject]