+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ---
 ## Master
 
+- Support for `optional` methods in ObjC protocols
 - Support for parsing lazy vars into Variable's attributes.
 - Updated Stencil to 0.13.1 and SwiftStencilKit to 2.7.0
 - In Swift templates CLI arguments should now be accessed via `argument` instead of `arguments`, to be consistent with Stencil and JS templates.